Add autosign-validator and more cucumber tests
The autosign-validator executable is the executable to be run by puppet
to validate autosigning.

The i-7672fe81.pem file is a certificate signing request generated by a
node with a CN of i-7672fe81 that has a corresponding secret of "secret"
to simplify testing. It has an extremely long (years) token expiration.