Remove safe_yaml gem
Remove dependencies on the safe_yaml gem because it has adverse
interactions with JVM puppetserver.

When safe_yaml is loaded into JVM puppetserver, it causes YAML fact
caches to be loaded as hashes rather than as serialized ruby objects,
which causes compiles to fail with an error about trusted facts that
reads:

Error 400 on SERVER: Attempt to assign to a reserved variable name:
'trusted' on node

There isn't a strong need for safe_yaml in autosign, because it's only
used to ensure safe loading of the configuration file. In principle,
somebody could now modify the config file to be a serialized ruby
object, but that person would probably already have sufficient access to
cause problems regardless.