Check PE autosign config file path
A recent change to puppet-autosign started putting config files for
Puppet Enterprise users in /etc/puppetlabs/puppetserver/autosign.conf,
but autosign doesn't actually check for a config file there.

This adds that as a path to check. My current understanding is that most
autosign users run PE, so I've put it in the first place to check as a
minor optimization.